Debian: Remove ARM naming quirk.
The machines had DNS entries for each network port as FOO-p0 and -p1, with FOO
being an alias. This meant Debian installer picked up the -p0 name for the
volume group. DNS has been fixed so that the primary host name omits the -p0
suffix and therefore matches e.g. the xenuse and osstest naming.