Debian: correct ARM u-boot boot runes
We currently need a comptible node of "xen,multiboot-module" as well as the
specific type. Also make sure the /chosen node specifies the number of address
and size-cells we will be using in the modules, otherwise Xen cannot parse
them.

Load Xen a bit higher up and use a variable so we don't need to update two
places.

Log some stuff as we go as well, so we can tell what was going on more easily.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>