target_file_exists: Use non-root access for this check
The new call of this in built_stash_file makes ts-*-build require root
access, which is not desirable.

The only other caller of target_file_exists is in ts-xen-install where
it is called on
                        /etc/default/xencommons
                        /etc/sysconfig/xencommons
                        /etc/default/xend
                        /etc/sysconfig/xend
all of which should be accessible as a normal user in a default config.

So change the access to use target_cmd_output rather than
target_cmd_output_root.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>