ts-debian-install: work around xen-create-image reliance on xend
The xen-tools script contains some sanity checks of the xend configuration and
fails if the configuration file is not present. This issue has been reported
upstream http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=732456.

Workaround this issue by creating a dummy xend-config.sxp if it does not
already exist. Include network- and vif- script options to keep the sanity
checks happy.

The workaround is enabled for all releases up to and including Wheezy in the
hopes that this will be fixed by the time Jessie is released, if not then the
osstest pushgate will hopefully catch this when Jessie is pushed to it.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>