cs-bisection-step: Use tree URLs from latest flight, not basis pass
We (both in adhoc-revtuple-generator and in the eventual build job)
(when using git) clone a particular url and then expect that the
resulting tree has all of the relevant revisions in it.

If a push gate has a different input tree to output tree, it is
necessary to use the output tree.  And, if the tree has moved, it is
probably more useful to use the new url than the old.

So it is more proper to use the tree url from the last flight found,
rather than that from the basis pass.  Do that.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>