make-flight: Adjust buildjob runvar for xend vs non-xend tests
Currently we only do so for xenbuildjob which only contains the hypervisor and
not the tools, for those we need buildjob. In principal the xenbuildjob output
should be identical for both xend and non-xend builds, but we use a matching
set of consistency.

This adds a xend suffix to buildjob for exactly test-amd64-i386-pv,
test-amd64-amd64-pv, test-amd64-i386-xend-winxpsp3,
test-amd64-i386-xend-qemut-winxpsp3 which is the expected set of xend jobs.