ts-kernel-build: Disable "Brocade BFA Fibre Channel Support"
This driver is broken on ARM:
    ERROR: "__bad_udelay" [drivers/scsi/bfa/bfa.ko] undefined!

I've taken the lazy way out and disabled it on all platforms. I think it isn't
especially likely that any of the current osstest hosts are using Fibre
Channel right now. The code to enable it came from a big batch addition of
drivers in 451f39c6149e.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
Acked-by: Ian Jackson <ian.jackson@eu.citrix.com>