Use ts-kernel-build for build-*-oldkern
instead of using the kernel build integration in xen.git, which is going away.

There is no difference to the .config produced. No test jobs seem to rely on
these kernels so I have not worried about making the contents of dist be
identical (specifically the filenames under /boot have lost their -xen suffix)

Remove the now unused xen and qemu(u) tree/revision stuff from the jobs'
runvars. Add the appropriate kconfighow and kimagefile runvars and implement
an appropriate kconfighow handler to use the create_config.sh present in this
tree (the xen.git intergration called the same script)

If the 2.6.18 build starts to fail, xen-unstable pushes will be broken
until they are fixed, as before.  Previously such a build failure might be due
to changes in xen.git so that was slightly justifiable.  But in practice this
is quite rare.  This situation, while anomalous, is tolerable.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
Acked-by: Ian Jackson <ian.jackson@eu.citrix.com>