distros: support PV guest install from Debian netinst media.
The netinst media are iso images containing a base Debian install and some
(image size dependent) additional tasks.

On x86 the "multiarch" iso flavour contains a Xen capable kernel for both i386
and amd64 so use that.

This adds support for two classes of ISO, the CD sized ones which are built
nightly and the DVD sized ones which are built weekly.

The images are downloaded using jigdo which sources the majority of the data
from a local Debian mirror, for this reason I have not worried about the fact
that the i386 and amd64 tests are downloading the same thing (adding a specific
download job would require finding up to 4GB of scratch space for each flight).

The ISOs booted using pygrub which can extract the kernel and initrd from a ISO
image. The resulting guests are also booted with pygrub since the pv-grub-menu
package is not available on the ISO images and we have pvgrub coverage from the
netboot tests.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>