Test pygrub and pvgrub on the regular flights
Since we now have the ability to test these drop one of each of pygrub,
pvgrub-32 and pvgrub-64 into the standard flights. Omitting the {Guest}_diver
runvar causes ts-debian-di-install to use the d-i images in the location
configured via TftpDiVersion, so they are version controlled along with the d-i
version used for the host.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>