Add a flight to test seabios.org's master branch
Most of this is the usual stuff for adding a new branch.

cr-daily-branch needed to set REVISION_SEABIOS for the seabios branch and leave
it blank for others (so that Config.mk's version will be used). ts-xen-build
already consumes the resulting version runvar but needed to be taught about the
tree URL runvar.

make-flight and mfi-common are updated to only build+test seabios on x86.

The new branch is added to cr-daily-branches.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
Acked-by: Ian Jackson <ian.jackson@eu.citrix.com>