rumpuserxen: Increase build timeout for grievous bodge
This involves copying 90Mby of data.  If the test machine is busy (in
the worst case, running the kernel side of resize2fs which due to
Linux's poor IO scheduling hogs the entire IO system) it can take a
long time.  In flight 26359 it was still running after 280s, when the
machine was released.

Increase the timeout to 2000s.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>