sg-run-job: Do blocked check globally, in run-job
Only run-job can fully abandon the whole job.  With the previous
arrangements (in bc63a901), a blocked build job would go on to try to
run ts-logs-capture and become broken (rather than blocked).

The downside of this is that it is not possible in the future to
introduce a kind of job which can selectively carry on if some of its
dependencies fail.  This notion is IMO an acceptable casualty.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>