make-flight: Do not filter the qemu-upstream-* jobs
When "qemu-xen-traditional" was the default, this filtering was
correct.  But now, "qemu-xen" is the default, and that version may
also be used in other tests for ancillary functions (eg, in dom0 for
the bootloader).

So simply do all of the jobs when testing qemu-upstream-*.

Retain the old behaviour for qemu-xen-4.2-testing, which has
qemu-xen-traditional as the default.  (There is no
qemu-xen-4.1-testing, and 4.3 has the change to the default.)

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>