executive-postgresql-schema: Schema dump
We really ought to have a proper schema script.  But for now, include
the output of pg_dump --schema-only on the osstestdb, for the benefit
of anyone trying to set up a hosted (rather than standalone) instance
(or just trying to see how the code works).

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>

---
v2: New patch in v2 of the series.