make-flight: rumpuserxen tests: use same buildjob for all guests
Change the name of the runvar for the build job to be
guests_rumpuserxenbuildjob, which is the default (fallback) for
guest_var.

(We are going to introduce another step which uses another guest but
wants the same build job.)

No resulting change in the system's overall behaviour.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>

---
v2: Fix stray leftover in commit message.