ts-guest-destroy-hard: New script
This just calls xl destroy; it doesn't check that the guest is
running, or on the network, or anything.  (It also doesn't check that
the guest is down afterwards.)

We are going to want this to be able to destroy the rumpuserxen
xenstorels domain, which is short-lived (and will normally have
terminated by the time we want to destroy it).

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>

---
v2: Fix copyright date in new script.