bisection: Do not get bisection start points from broken flights
cs-bisection-step ignores any flight with broken steps, on the grounds
that the results in such a flight maybe unreliable.

Therefore, if a main flight has both broken steps and regressions, it
is wrong to take its mro as the new basis mro for bisections.  In the
worst case a newly failing job will be determined to be unbisectable
(since the only relevant failure is in a broken flight), have the
stamp file created, and then never be properly bisected.

So: in sg-report-flight, report `broken-step' for each broken step (in
the top-level flight we are reporting - broken jobs in flights we are
referring to to justify heisenbugs are OK); and, in cr-daily-branch,
do not install the new mro if such indications are present.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>