ts-hosts-allocate-Executive: Score for equivalent previous failures
Look to see whether the last run on any hosts which are equivalent to
the ones we're looking at, failed.  This means that when host X is
failing and we are considering host Y which is equivalent to X, we
give Y a selection bonus.

This means that osstest will be less obsessive about sticking to the
very same failing host.

Also, no longer multiply by the number of previously-failing hosts;
instead, just consider whether _all_ the hosts are previously-failing
in their corresponding slots (*_host runvars).  It doesn't make sense
to make the bonus for using failing hosts depend on the number of
hosts used by the test.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>

---
v2: Improve commit message.