ts-guest-start: Use guest_create
This allows us to abolish CfgPathVar which was inconsistently used,
appears redundant with $gho->{CfgPath} and in any case never set to
anything other than 'cfgpath'.

This was intended to deal with toolstacks with a different cfg format
(such as very early versions of xl) but is no longer relevant.

I think if this need arises again in a future toolstack this
functionality could be reintroduced pretty trivially via the toolstack
abstraction which is added by this series.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
Acked-by: Ian Jackson <ian.jackson@eu.citrix.com>